+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Smelte changelog
 
+# 0.1.6
+* Add/replace/remove to chips
+* Add anchor to buttons
+
 # 0.1.5
 * Fix jerkiness in expanding treeview
 * Add code example to treeview